This function estimates the fractional occupancy or probability of a series of states.'

This function estimates the fractional occupancy or probability of a series of states.'

Usage

nest_fo(tbl, vars, foVar)

Arguments

tbl

A tibble/dataframe featuring a list of states in time - and covariates

vars

Variables of interest in the tibble/dataframe

foVar

States variable name to obtain fractional occupancy.